init backend - just print hello world
