+++ README.md
@@ -14,6 +14,10 @@ pnpm dev
 bun dev
 ```
 
+# To run python:
+python -m venv venv
+.\venv\Scripts\activate
+
 Open [http://localhost:3000](http://localhost:3000) with your browser to see the result.
 
 You can start editing the page by modifying `app/page.tsx`. The page auto-updates as you edit the file.
